Refactor database schema: consolidate constants into a new file, remove supervisor schema (now it's a globalRole), and update club and event relations. Introduce clubMembership and eventRegistration tables with validation schemas.
