+++ db/schema/index.ts
@@ -2,6 +2,7 @@ export * from './user';
 export * from './club';
 export * from './event';
 export * from './ticket';
-export * from './supervisor';
 export * from './relations';
 export * from './task';
+export * from './clubMembership';
+export * from './eventRegistration';
